App/requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
docutils==0.20.1
gitdb==4.0.11
GitPython==3.1.40
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Kivy==2.2.1
Kivy-Garden==0.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
postgrest==0.15.0
protobuf==4.25.1
pyarrow==14.0.1
pycairo @ file:///private/tmp/py3cairo-20231021-6574-xwz2b1/pycairo-1.25.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.46.0
python-dateutil==2.8.2
pytz==2023.3.post1
realtime==1.0.2
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
st-supabase-connection==1.2.2
storage3==0.7.0
streamlit==1.28.2
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
websockets==11.0.3
zipp==3.17.0
psycopg2-binary==3.1
sqlalchemy==2.0.25